update memeCount test
